Added inline css for style attributes
Added more patterns to meta tag section. This adds css syntax hightlighting to style attributes on DOM elements.
Known issue: due to the end capture used in css attribute lists the syntax hightlighting is sensitive to closing values with a semicolon. It looks like the css rule-list-innards grammar uses ; or } to end the capture.